requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1

snowflake-connector-python[pandas]==3.10.1
snowflake-snowpark-python==1.37.0

numpy==1.26.4
pandas==2.1.4
pyarrow==15.0.2

starlette==0.37.2
pydantic==2.8.2
pydantic-core==2.20.1
typing-extensions==4.12.2
anyio==4.4.0

requests==2.31.0
tinydb==4.8.0
plotly==5.22.0
dash==2.17.1
